backend-python/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
pymongo==4.4.1
motor==3.2.0
pydantic==2.2.0
requests==2.31.0
PyJWT==2.8.0
werkzeug==2.3.7
gunicorn==21.2.0
reportlab==4.0.4
python-dateutil==2.8.2
pytz==2023.3
